Fix tests

+++ api/tests/utils_db.py
@@ -1,16 +1,12 @@
 import os
-from typing import List
-
 import pytest
-
-from src.models import Article, Category, User, LoginLog, ExerciseCoherenceCardiac  # noqa: F401
-from sqlmodel import create_engine
+from typing import List
+from sqlmodel import SQLModel, create_engine
+from sqlmodel.ext.asyncio.session import AsyncSession
 from sqlalchemy.ext.asyncio import AsyncEngine, create_async_engine
 from sqlalchemy.orm import sessionmaker
-from sqlmodel.ext.asyncio.session import AsyncSession
+from src.models import Article, Category, User, LoginLog, ExerciseCoherenceCardiac  # noqa: F401
 
-
-from sqlmodel import SQLModel
 
 IS_ECHO = True
 DB_NAME = "test.db"
@@ -36,17 +32,18 @@ async def get_test_session() -> AsyncSession:
 
 
 @pytest.fixture(autouse=True)
-def reset_test_db()->None:
-    sep_lines = "_"*50
-    print(f"Reset db...\n{sep_lines}")
-    os.remove(DB_NAME)
-    print(f"Done!\n{sep_lines}")
+def reset_test_db() -> None:
+    sep_lines = "_" * 50
+    if os.path.isfile(DB_NAME):
+        print(f"Reset db...\n{sep_lines}")
+        os.remove(DB_NAME)
+        print(f"Done!\n{sep_lines}")
     print(f"Load entities...\n{sep_lines}")
     SQLModel.metadata.create_all(sqlite_sync_engine)
     print(f"Done!\n{sep_lines}")
 
 
-async def load_objects(objects: List[SQLModel])->None:
+async def load_objects(objects: List[SQLModel]) -> None:
     async with AsyncSession(
         sqlite_async_engine,
         expire_on_commit=False,
